requirements.txt
# Core web + database (STEP 1)
Flask>=3.0,<4
Flask-SQLAlchemy>=3.1,<4
Flask-Migrate>=4.0,<5
SQLAlchemy>=2.0,<3
psycopg2-binary>=2.9,<3
python-dotenv>=1.0,<2

# Used in later steps (auth)
Flask-Login>=0.6,<1
PyJWT>=2.8,<3

# Production WSGI server (Render, etc.) — required for the Procfile "web" command
gunicorn>=23.0

# Tests (pytest + Flask test client)
pytest>=8.0,<9
